add/fix manual browser script
