+++ package.json
@@ -19,7 +19,7 @@
   "devDependencies": {
     "@rollup/plugin-commonjs": "^23.0.2",
     "@rollup/plugin-node-resolve": "^15.0.1",
-    "rollup": "^3.2.5",
-    "rollup-plugin-uglify": "^6.0.4"
+    "@rollup/plugin-terser": "^0.1.0",
+    "rollup": "^3.2.5"
   }
 }
